+++ common/include/common/config/GameConfig.h
@@ -32,8 +32,8 @@ struct GameConfig
 
     static constexpr unsigned min_fps_limit = 10u;
     static constexpr unsigned max_fps_limit = 240u;
-    CfgVar<unsigned> max_fps{60, [](auto val) { return std::clamp(val, min_fps_limit, max_fps_limit); }};
-    CfgVar<unsigned> server_max_fps{60, [](auto val) { return std::clamp(val, min_fps_limit, max_fps_limit); }};
+    CfgVar<unsigned> max_fps{120, [](auto val) { return std::clamp(val, min_fps_limit, max_fps_limit); }};
+    CfgVar<unsigned> server_max_fps{120, [](auto val) { return std::clamp(val, min_fps_limit, max_fps_limit); }};
 
     // Graphics
     CfgVar<bool> fast_anims = false;