+++ docs/CHANGELOG.md
@@ -51,6 +51,7 @@ Version 1.8.0 (not released yet)
 - Fix crash when skipping a cutscene that stared when blackout was in progress
 - Remove limit of restored objects during a level transition
 - Fix possible buffer overflow if transitioned object mesh name is longer than 31 characters
+- Change default FPS limit to 120 to comply with the stock game
 
 Version 1.7.0 (released 2021-06-05)
 -----------------------------------